refactor: ApplyCouponPolicyUseCase 이동 및 관련 data Domain으로 이동
